Remove shadow analyzer from TOOLS_NOGO (#4344)
**What type of PR is this?**

Other

**What does this PR do? Why is it needed?**

The shadow analyzer issues many false positives which are not in keeping
with Go best practices. Even the upstream documentation for this
analyzer admits this and claims that it is experimental. It should not
be included in the default list of analyzers provided by nogo (users can
still opt-in to it if they want it).


**Which issues(s) does this PR fix?**

Fixes #4340

**Other notes for review**