Drop non-hermetic deps in _go_tool_binary_impl (#4365)
**What type of PR is this?**
Bug fix for Nix

**What does this PR do? Why is it needed?**
The current implementation depends on `mktemp` and `rm` being in the
PATH, which they're not on Nix. This alternative construction executes
`go build` directly without run_shell or the coreutils deps.

While here, I set a few extra env vars to disable new Go features we
don't want (GOTELEMETRY, GOENV).

If this works, we may be able to remove the Windows codepath.

**Which issues(s) does this PR fix?**

Fixes #

**Other notes for review**

Co-authored-by: Jay Conrod <jay@engflow.com>