Request stdlib cache dir in gopackagesdriver (#4391)
<!-- Thanks for sending a PR! Before submitting:

1. If this is your first PR, please read CONTRIBUTING.md and sign the
CLA
   first. We cannot review code without a signed CLA.
2. Please file an issue *first*. All features and most bug fixes should
have
an associated issue with a design discussed and decided upon. Small bug
   fixes and documentation improvements don't need issues.
3. New features and bug fixes must have tests. Documentation may need to
be updated. If you're unsure what to update, send the PR, and we'll
discuss
   in review.
4. Note that PRs updating dependencies and new Go versions are not
accepted.
   Please file an issue instead.
-->

**What type of PR is this?**

Bug fix

**What does this PR do? Why is it needed?**

The stdlib_json_file contains paths of compiled cgo files, which are in
the stdlib cache_dir. The gopackagesdriver reads those files, but this
dependency was not explicitly declared previously.

Usually, it works despite the missing dependency. But when you have a
Bazel cache configured, and then run `bazel clean` and run the
gopackagesdriver again, Bazel will not restore the cache_dir from the
Bazel cache. The gopackagesdriver then fails with an error like this:

error: unable to load JSON files: unable to resolve imports: open
[...]/bin/external/rules_go+/stdlib_/gocache/a4/a4f91e8314b27a4bce0e8fbc01bb9736cb9ed10da747af722c5a4f2fcb1213a3-d:
no such file or directory

This change adds a new output group containing the stdlib cache_dir, and
requests this from the gopackagesdriver, fixing the problem.