wrap_sdk: make root_files a string_keyed_label_dict (#4419)
**What type of PR is this?**

Bug fix

**What does this PR do? Why is it needed?**
I'm trying to use an extension-generated repo as a wrapped sdk using
`root_files`. I get a package load error similar to
https://github.com/bazelbuild/bazel/issues/19301.

**Which issues(s) does this PR fix?**
I saw in the comments (specifically
https://github.com/bazelbuild/bazel/issues/19301#issuecomment-1694588371)
that this would probably work if `root_files` used a
`string_keyed_label_dict` and, when I tried it locally it did.

Fixes #4418

**Other notes for review**
https://github.com/bazelbuild/bazel/issues/19301