Revert back from attribute to rule transitions (#4415)
**What type of PR is this?**

Refactoring

**What does this PR do? Why is it needed?**

Partially reverts #3116, which is no longer necessary now that Bazel 7
and higher support configurable attributes with rule transitions. As a
result, a top-level Go rule shares its configuration with that of the
go_context and cgo_context targets it depend on. This unblocks switching
to joint toolchain resolution for Go and C/C++ in a backwards-compatible
way.

**Which issues(s) does this PR fix?**

**Other notes for review**