Fix trim path problems (#4435)
**What type of PR is this?**

Bug fix

**What does this PR do? Why is it needed?**

There were multiple problems with trim paths:

- The trim path was inconsistent between non-cgo and cgo. non-cgo
trimmed the working directory, and cgo trimmed the parent of the working
directory.
- For non-cgo, paths in external repos are not trimmed when
--experimental_sibling_repository_layout is enabled.
- For cgo, the name of the working directory is included in the output,
but this name is not consistent between local and remote builds,
breaking reproducibility.

These are fixed here by creating the trim path in the same way for both
non-cgo and cgo, and trimming the working directory as well as replacing
the parent of the working directory with `..`. This means that the
resulting paths are relative to the working directory.

Additionally, this fixes the existing feature to extend an existing
trimpath. It used the wrong separator (`:` instead of `;`), and it
applied `abs()` to the entire trimpath argument instead of the
individual paths inside it. (See
https://github.com/bazel-contrib/rules_go/pull/2994#issuecomment-1012730010
and
https://github.com/bazel-contrib/rules_go/pull/2994#issuecomment-1012796895)

**Which issues(s) does this PR fix?**

Fixes: https://github.com/bazel-contrib/rules_go/issues/4434
Fixes: https://github.com/bazel-contrib/rules_go/issues/4161